Add an additional API function hook, and separate hooks into protected (required) and unprotected (optional).

+++ Source/HookModuleMain.cpp
@@ -20,8 +20,11 @@
 
 // -------- MACROS --------------------------------------------------------- //
 
-/// Convenience wrapper for instantiating a hook record structure for the given named Windows API function.
-#define HOOK_RECORD(windowsApiFunctionName)                                 {.setHookFunc = ::Pathwinder::Hooks::ProtectedDependency::##windowsApiFunctionName::SetHook, .hookProxy = DynamicHook_##windowsApiFunctionName::GetProxy()}
+/// Convenience wrapper for instantiating a hook record structure for a protected hook the given named Windows API function.
+#define PROTECTED_HOOK_RECORD(windowsApiFunctionName)                       {.protectedHookSetFunc = ::Pathwinder::Hooks::ProtectedDependency::##windowsApiFunctionName::SetHook, .hookProxy = DynamicHook_##windowsApiFunctionName::GetProxy()}
+
+/// Convenience wrapper for instantiating a hook record structure for an unprotected hook given the named Windows API function.
+#define UNPROTECTED_HOOK_RECORD(windowsApiFunctionName)                     {.unprotectedHookOriginalAddress = GetInternalWindowsApiFunctionAddress(#windowsApiFunctionName), .hookProxy = DynamicHook_##windowsApiFunctionName::GetProxy()}
 
 
 namespace Pathwinder
@@ -33,7 +36,8 @@ namespace Pathwinder
         /// Holds together all of the information needed to attempt to set a Hookshot dynamic hook.
         struct DynamicHookRecord
         {
-            Hookshot::EResult(*setHookFunc)(Hookshot::IHookshot*);          ///< Address of the function that will be invoked to set the hook.
+            Hookshot::EResult(*protectedHookSetFunc)(Hookshot::IHookshot*); ///< Address of the function that will be invoked to set the hook. Used by protected hooks only, as unprotected hooks are set via the hook proxy.
+            void* unprotectedHookOriginalAddress;                           ///< Address of the original function to be hooked. Used by unprotected hooks only and passed to the hook proxy.
             Hookshot::DynamicHookProxy hookProxy;                           ///< Proxy object for the dynamic hook object itself.
         };
 
@@ -45,24 +49,44 @@ namespace Pathwinder
         {
             // References the hooks declared in "Hooks.h" and must contain all of them.
             const DynamicHookRecord hookRecords[] = {
-                HOOK_RECORD(NtClose),
-                HOOK_RECORD(NtCreateFile),
-                HOOK_RECORD(NtOpenFile),
-                HOOK_RECORD(NtQueryDirectoryFile),
-                HOOK_RECORD(NtQueryDirectoryFileEx),
-                HOOK_RECORD(NtQueryInformationByName)
+                PROTECTED_HOOK_RECORD(NtClose),
+                PROTECTED_HOOK_RECORD(NtCreateFile),
+                PROTECTED_HOOK_RECORD(NtOpenFile),
+                PROTECTED_HOOK_RECORD(NtQueryDirectoryFile),
+                PROTECTED_HOOK_RECORD(NtQueryDirectoryFileEx),
+                PROTECTED_HOOK_RECORD(NtQueryInformationByName),
+                UNPROTECTED_HOOK_RECORD(NtQueryAttributesFile)
             };
 
             Message::OutputFormatted(Message::ESeverity::Debug, L"Attempting to hook %u Windows API function(s).", (unsigned int)_countof(hookRecords));
 
             for (const auto& hookRecord : hookRecords)
             {
-                const Hookshot::EResult setHookResult = hookRecord.setHookFunc(hookshot);
-
-                if (false == Hookshot::SuccessfulResult(setHookResult))
+                if (nullptr != hookRecord.protectedHookSetFunc)
                 {
-                    Message::OutputFormatted(Message::ESeverity::ForcedInteractiveError, L"%.*s failed to set a hook for the Windows API function \"%s\" and cannot function without it.\n\nHookshot::EResult = %u", (int)Strings::kStrProductName.length(), Strings::kStrProductName.data(), hookRecord.hookProxy.GetFunctionName(), (unsigned int)setHookResult);
-                    TerminateProcess(Globals::GetCurrentProcessHandle(), (UINT)-1);
+                    // Hook is protected.
+                    // Setting it must succeed in order for Pathwinder to function correctly.
+
+                    const Hookshot::EResult setHookResult = hookRecord.protectedHookSetFunc(hookshot);
+
+                    if (false == Hookshot::SuccessfulResult(setHookResult))
+                    {
+                        Message::OutputFormatted(Message::ESeverity::ForcedInteractiveError, L"%.*s failed to set a hook for the Windows API function \"%s\" and cannot function without it.\n\nHookshot::EResult = %u", (int)Strings::kStrProductName.length(), Strings::kStrProductName.data(), hookRecord.hookProxy.GetFunctionName(), (unsigned int)setHookResult);
+                        TerminateProcess(Globals::GetCurrentProcessHandle(), (UINT)-1);
+                    }
+                }
+                else
+                {
+                    // Hook is unprotected.
+                    // Setting it successfully is considered optional.
+
+                    const Hookshot::EResult setHookResult = hookRecord.hookProxy.SetHook(hookshot, hookRecord.unprotectedHookOriginalAddress);
+
+                    if (false == Hookshot::SuccessfulResult(setHookResult))
+                    {
+                        Message::OutputFormatted(Message::ESeverity::Warning, L"Failed to hook the \"%s\" Windows API function (Hookshot::EResult = %u).", hookRecord.hookProxy.GetFunctionName(), static_cast<unsigned int>(setHookResult));
+                        continue;
+                    }
                 }
 
                 Message::OutputFormatted(Message::ESeverity::Debug, L"Successfully hooked the \"%s\" Windows API function.", hookRecord.hookProxy.GetFunctionName());
